Move array insertion util to util namespace
